+++ dev/library/library/src/main/java/com/kodilla/library/repository/TitleDao.java
@@ -1,9 +1,7 @@
 package com.kodilla.library.repository;
 
-import com.kodilla.library.domain.Copy;
 import com.kodilla.library.domain.Title;
 import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +9,6 @@ import java.util.List;
 
 @Transactional
 @Repository
-public interface TitleDao extends CrudRepository<Title,Integer> {
+public interface TitleDao extends CrudRepository<Title,Long> {
 
 }
